Support for very large book files. Also fixes https://github.com/syzygy1/Cfish/issues/58

+++ src/tt.c
@@ -197,13 +197,13 @@ TTEntry *tt_probe(Key key, int *found)
   for (int i = 0; i < ClusterSize; i++)
     if (!tte[i].key16 || tte[i].key16 == key16) {
       if ((tte[i].genBound8 & 0xFC) != TT.generation8 && tte[i].key16)
-        tte[i].genBound8 = (uint8_t)(TT.generation8 | tte_bound(&tte[i])); // Refresh
-      *found = (int)tte[i].key16;
+        tte[i].genBound8 = TT.generation8 | tte_bound(&tte[i]); // Refresh
+      *found = tte[i].key16;
       return &tte[i];
     }
 
   // Find an entry to be replaced according to the replacement strategy
-  TTEntry* replace = tte;
+  TTEntry *replace = tte;
   for (int i = 1; i < ClusterSize; i++)
     // Due to our packed storage format for generation and its cyclic
     // nature we add 259 (256 is the modulus plus 3 to keep the lowest
@@ -232,4 +232,3 @@ int tt_hashfull(void)
   }
   return cnt;
 }
-
